docs: 일부 주석 처리 및 파일 경로 수정
